+++ doc/ReleaseNotes.md
@@ -1,6 +1,17 @@
 ## New in v1.29
 
-Nothing yet.
+
+### Manifest `<PACKAGEVERSION>` token expansion
+
+Added `<PACKAGEVERSION>` token which will use the `PackageVersion` from the manifest in place of the token. This applies to the following manifest fields:
+
+* `NestedInstallerFiles.RelativeFilePath`
+* `ProductCode`
+* `AppsAndFeaturesEntries.DisplayName`
+* `AppsAndFeaturesEntries.ProductCode`
+* `InstallationMetadata.DefaultInstallLocation`
+* `InstallationMetadata.Files.RelativeFilePath`
+* `ReleaseNotesUrl`
 
 ## Bug Fixes